Find framework include paths on the Mac.
The pp-engine couldn't handle include commands that refer to headers in frameworks on Mac.
If frameworks are specified to the compiler/preprocessor, includes like "Somepath/myfile.h" are
also broken down to "Somepath.framework/Headers/myfile.h" in the framework search paths.
The standard Qt binary install on the Mac is compiled with frameworks, so this is of some relevance.